Conflicts:
	pymotifs/loops/extractor.py

+++ pymotifs/loops/positions.py
@@ -5,31 +5,52 @@ from pymotifs import core
 from pymotifs import utils
 from pymotifs.models import LoopPositions
 from pymotifs.models import LoopsAll
+from pymotifs.loops.extractor import Loader as InfoLoader
 
 
-class Loader(core.SimpleLoader):
+class Loader(core.Loader):
     merge_data = True
+    dependencies = set([InfoLoader])
 
     def __init__(self, *args, **kwargs):
         super(Loader, self).__init__(*args, **kwargs)
-        self.matlab = core.Matlab(self.config['locations']['fr3d_root'])
         self.precomputed = self.config['locations']['loops_mat_files']
-
-    def parse(self, filename):
-        with open(filename, 'rb') as raw:
-            reader = csv.reader(raw, delimiter=',', quotechar='"')
-            for row in reader:
-                yield row
-
-    def has_data(self, pdb, **kwargs):
-        return False
 
     def remove(self, pdb, **kwargs):
         with self.session() as session:
-            session.query(LoopPositions).\
+            query = session.query(LoopsAll).filter_by(pdb=pdb)
+            ids = [result.id for result in query]
+
+        if not ids:
+            return True
+
+        with self.session() as session:
+            return session.query(LoopPositions).\
+                filter(LoopPositions.loop_id.in_(ids)).\
+                delete(synchronize_session=False)
+
+        return True
+
+    def has_data(self, pdb, **kwargs):
+        with self.session() as session:
+            query = session.query(LoopPositions).\
                 join(LoopsAll, LoopsAll.id == LoopPositions.loop_id).\
-                filter(LoopsAll.pdb == pdb).\
-                delete(synchronize_session='fetch')
+                filter(LoopsAll.pdb == pdb)
+            return bool(query.count())
+
+    def parse(self, filename):
+        keys = ['loop_id', 'position', 'unit_id', 'bulge', 'flanking',
+                'border']
+        int_keys = ['position', 'bulge', 'flanking', 'border']
+        data = []
+        with open(filename, 'rb') as raw:
+            reader = csv.reader(raw, delimiter=',', quotechar='"')
+            for row in reader:
+                entry = dict(zip(keys, row))
+                for key in int_keys:
+                    entry[key] = int(entry[key])
+                data.append(entry)
+            return data
 
     def known(self, pdb):
         mapping = {}
@@ -44,37 +65,37 @@ class Loader(core.SimpleLoader):
 
         return mapping
 
-    def get_annotations(self, pdb):
-        path = os.path.join(self.precomputedData, pdb)
-        [output_file, err_msg] = self.mlab.loadLoopPositions(path, nout=2)
+    def annotation_file(self, pdb):
+        matlab = core.Matlab(self.config['locations']['fr3d_root'])
+        path = str(os.path.join(self.precomputed, pdb))
+        if not os.path.exists(path):
+            os.mkdir(path)
+
+        [output_file, err_msg] = matlab.loadLoopPositions(path, nout=2)
         if err_msg != '':
             raise core.MatlabFailed(err_msg)
 
         return output_file
 
-    def data(self, pdb):
+    def annotations(self, pdb, remove=True):
+        output_file = self.annotation_file(pdb)
+        data = self.parse(output_file)
+        if remove:
+            os.remove(output_file)
+        return data
+
+    def data(self, pdb, **kwargs):
         """Update loop_positions table by loading data from the mat files
         stored in the PrecomputedData folder
         """
         data = []
-        output_file = self.get_annotations(pdb)
         known = self.known(pdb)
-        for row in self.parse(output_file):
-            (loop_id, position, nt_id, bulge, flanking, border) = row
-            entry = known.get((loop_id, position), {})
+        for row in self.annotations(pdb):
+            entry = known.get((row['loop_id'], row['position']), {})
             if not entry:
-                self.logger.info("New loop %s found", loop_id)
+                self.logger.info("New loop %s found", row['loop_id'])
 
-            entry.update({
-                'loop_id': loop_id,
-                'position': position,
-                'nt_id': nt_id,
-                'flanking': int(flanking),
-                'bulge': int(bulge),
-                'border': int(border)
-            })
+            entry.update(row)
             data.append(LoopPositions(**entry))
-
-        os.remove(output_file)  # delete temporary csv file
 
         return data
